BTCARDSXP-60 Add support for creating transactions with line items (#156)
What
===
Add support for including line items in transaction create and handling
their validation errors.

Why
===
Line items have been added to the gateway and we want merchants to be
able to use the server SDKs to include line items for a transaction when
submitting a transaction.

Note
===
There is no validating that line items are being returned because a
transaction create response doesn't include the line items for the
transaction. There will be a separate change made to add support for
getting line items from a separate endpoint.