BTCARDSXP-60 Add support for getting line items for a transaction (#157)
What
===
Add support for getting line items for a transaction.

Why
===
Line items are created when creating a transaction but they aren't
returned in the transaction response, so the gateway exposes line items
for symmetry on a seprate endpoint. We should add it to the server SDKs
so merchant can access line items they've stored on transactions.