BTCARDSXP-60 Add tests for every line items validation error (#159)
What
===
Add tests for every validation error for the line item fields, where
that validation error can be triggered via incorrect data being given to
the Braintree Java SDK.

Why
===
Tests for validation errors help teach others how the validation errors
can occur. This feedback was given in another PR so we're going back and
adding them for these validation errors that were just recently added.

Note
===
Tests were not added for validation errors that are not possible to
encounter with the Java SDK because of type safety, namely the invalid
format validation errors that are prevented because you can't create
a `BigDecimal` with letters.