BTCOMINF-305 - Add `signupOnly` option to `OAuthConnectUrlRequest`s
Magento wanted a way to force signup only on OAuth connections. Work has
been done in Signups (see
<https://github.braintreeps.com/braintree/signups/pull/430>) to allow
this, and now we need to add a flag to generate the connect URL
properly.

The only hairy piece was handling cases where both `loginOnly` and
`signupOnly` are true. Because I wanted to preserve backwards
compatibility, I opted to never throw errors and let `loginOnly` "win"
if both were set. If backwards compatibility weren't a concern, I'd make
these throw errors.

This is a private feature for Magento, which is why I didn't put
anything in the changelog, and why we don't intend to add it to any of
the other server SDKs.