BTCARDSXP-952 Remove .rvmrc (#222)
### What
Remove .rvmrc that installs the latest available version 2 of ruby when
entering the directory.

### Why
The .rvmrc is more annoying than useful. At the moment it is attempting
to install ruby `2.6.0-preview2`, which isn't even a stable version.

Most of the SDKs do not have a .rvmrc. The only use of ruby in this repo
is for the Rakefile. We mostly use the Makefile and Dockerfile to setup
the environment for running tests locally, and if we want a specific
version of ruby it needs to be set or changed in there since the .rvmrc
is not used inside the Docker container.