[PayPal] Billing Address For Payer Info
We at the New York Times are in need of retrieving the billing address
for a given Billing Agreement Nonce using the Vault Flow.

We would like to propose the follow changes to satisfy this need.

Changes

- Adds Address object to enable retrieval of the PayPal Billing Address
from the backend.

Side Effects

Can expect the Billing Address to be populated from the Billing Address
found in PayPal.