DTBTFTAUTH - Fix integration test for 3ds
Why This Change Is Necessary
========================================================================

Fix broken build

What These Changes Do To Address the Issue
========================================================================

During this integration test, the gateway is sending a request to
Cardinal, and cardinal is not returning `liability_shift` value of true
if the card we use for the test transaction is expired.  Changed
expiration date from `2020-12` to `2030-12`, so we are ok for another 10
years.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

Signed-off-by: Mami Nomura <mami.nomura@getbraintree.com>
Signed-off-by: Alejandro Desatnik Sod <alejandro.desatniksod@getbraintree.com>
Signed-off-by: Mami Nomura <mami.nomura@getbraintree.com>