Fix merchant account tests using deprecated create feature Ref: https://braintree.slack.com/archives/C0MPVCMGW/p1648659334362859?thread_ts=1648653431.290839&cid=C0MPVCMGW
the Create route had the surfacing of multiple validation errors removed
via a PFS in both sand & prod previously.  We've since been testing a
removed feature in the sdk up until the dead code was removed.
Adjusting this to use update restores the scope of the test to a live
code path/behavior matching sand/prod.