DTBTSDK-1682 FindAllStrings safely handles string OR regex arguments (#431)
We received a [GH issue][1] that certain search requests were failing once
they updated to v 3.16.0. On closer inspection, there was a code change
to resolve an issue where `findAllStrings` was returning children XML
nodes as one giant string. Unfortunately the fix for this issue then
broke the ResourceCollection class that used `findAllStrings` with a
regex expression ("ids/*") as its argument.

Beefs up our unit tests so that we're testing the different ways
`findAllStrings` can be used.
Fixes findAllStrings to safely handle scenarios that may not be using a
regex expression as the argument.

[1]: https://github.com/braintree/braintree_java/issues/108
JIRA: https://engineering.paypalcorp.com/jira/browse/DTBTSDK-1682

Signed-off-by: Holly Stotelmyer <holly.stotelmyer@getbraintree.com>
Co-authored-by: Debra Do <code@getbraintree.com>