DTBTCRDSXP-2462 recurring_first should not be final auths (#448)
What:
Change final auths test setup to use recurring transactions instead of
recurring_first.

Why:
The current test setup relies on a bug in our code. Recurring
transactions are merchant initiated, and truly final auths, as in no
adjustments allowed. Recurring first are customer initiated and hence
it's fine to allow adjustments for these.

We plan on fixing this in the gateway, and this is a pre-requisite so
that the SDK integration builds don't break.